Clean up parser
