+++ src/lib/parser/index.js
@@ -1,0 +1,7 @@
+'use strict';
+function __export(m) {
+  for (var p in m) if (!exports.hasOwnProperty(p)) exports[p] = m[p];
+}
+Object.defineProperty(exports, '__esModule', { value: true });
+__export(require('./fromBuffer'));
+__export(require('./toBuffer'));
