+++ src/lib/parser/toBuffer.js
@@ -1,0 +1,6 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+function psbtToBuffer(psbtData) {
+  return Buffer.from([psbtData.inputs.length]);
+}
+exports.psbtToBuffer = psbtToBuffer;
