+++ ts_src/lib/parser/toBuffer.ts
@@ -1,0 +1,5 @@
+import { PsbtAttributes } from './index';
+
+export function psbtToBuffer(psbtData: PsbtAttributes): Buffer {
+  return Buffer.from([psbtData.inputs.length]);
+}
