+++ types/lib/parser/fromBuffer.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="node" />
+import { PsbtAttributes } from './index';
+export declare function psbtFromBuffer(buffer: Buffer): PsbtAttributes;
