+++ types/lib/parser/index.d.ts
@@ -1,0 +1,10 @@
+import { Transaction } from 'bitcoinjs-lib';
+import { PsbtGlobal, PsbtInput, PsbtOutput } from '../interfaces';
+export * from './fromBuffer';
+export * from './toBuffer';
+export interface PsbtAttributes {
+    unsignedTx: Transaction;
+    globalMap: PsbtGlobal;
+    inputs: PsbtInput[];
+    outputs: PsbtOutput[];
+}
