+++ types/lib/parser/toBuffer.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="node" />
+import { PsbtAttributes } from './index';
+export declare function psbtToBuffer(psbtData: PsbtAttributes): Buffer;
